[#5][#8]Fix error checking overwriting And increments File Output
Fixes #5, #8

Error checking will replace rows 1 at a time, as completed.

File output will check to see if the output file already exists
and increment the number by 1 if so, e.g. if 'test_results.csv'
is already there, it will make the output file 'test_results(1).csv'
and if that already exists, it will be 'test_results(2).csv' and
so and so forth...